refactor: reorganize qrcode module
