+++ packages/cryptoplease/lib/features/qr_scanner/module.dart
@@ -1,2 +1,2 @@
-export 'src/qr_address_data.dart';
-export 'src/qr_scanner_request.dart';
+export 'src/bl/qr_address_data.dart';
+export 'src/bl/qr_scanner_request.dart';
