+++ packages/cryptoplease/lib/features/qr_scanner/src/bl/qr_address_data.dart
