+++ packages/cryptoplease/lib/features/qr_scanner/src/bl/qr_scanner_request.dart
@@ -3,7 +3,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:solana/solana.dart';
 import 'package:solana/solana_pay.dart';
 
-import 'qr_address_data.dart';
+import '../bl/qr_address_data.dart';
 
 part 'qr_scanner_request.freezed.dart';
