+++ packages/cryptoplease/lib/features/qr_scanner/src/bl/qr_scanner_state.dart
