+++ packages/cryptoplease/lib/features/qr_scanner/src/presentation/qr_scanner_background.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../../../gen/assets.gen.dart';
+import '../../../../gen/assets.gen.dart';
 
 class QrScannerBackground extends StatelessWidget {
   const QrScannerBackground({
