+++ packages/cryptoplease/lib/features/qr_scanner/src/presentation/qr_scanner_screen.dart
@@ -6,13 +6,13 @@ import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:mobile_scanner/mobile_scanner.dart';
 
-import '../../../di.dart';
-import '../../../gen/assets.gen.dart';
-import '../../../l10n/l10n.dart';
-import '../../../ui/dialogs.dart';
-import '../../../ui/theme.dart';
+import '../../../../di.dart';
+import '../../../../gen/assets.gen.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/dialogs.dart';
+import '../../../../ui/theme.dart';
+import '../bl/qr_scanner_bloc.dart';
 import 'qr_scanner_background.dart';
-import 'qr_scanner_bloc.dart';
 
 class QrScannerScreen extends StatelessWidget {
   const QrScannerScreen({Key? key}) : super(key: key);
